requirements.txt
Django==3.1
django-tables2
django-crispy-forms
django-simple-menu
django-calculation
django-extra-views
